// src/components/solicitudes-pago/solicitudes-pago.types.tsx
export type CrearSolicitudProveedorPayload = {
  tipo_solicitud: "PAGO_PROVEEDOR";
  proyecto_base_id: string;
  centro_costo_id: string;
  beneficiario_id: string;
  categoria_gasto: string;
  medio_pago: MedioPagoSolicitud;
  descripcion: string;
  valor_bruto: number;
  valor_impuestos: number;
  valor_retenciones: number;
  valor_descuentos: number;
};